requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
distlib==0.3.6
django==3.2.0
djangorestframework==3.12.4
filelock==3.10.7
identify==2.5.22
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pluggy==0.13.1
pre-commit==3.2.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.2
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tzdata==2023.2
urllib3==1.26.15
virtualenv==20.21.0
python-dotenv==1.0.0
